Add functionality to remove book and dispatch action
