--- src/Components/categoriesButton.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const CategoryButton = () => <button type="button">Check status</button>;
-
-export default CategoryButton;
