--- src/Pages/Error.js
@@ -1,3 +1,0 @@
-export default function NotFound() {
-  return <div>Not found</div>;
-}
