--- src/redux/shared/actions.js
@@ -1,3 +1,0 @@
-const RECEIVE_DATA = 'bookstore/shared/RECEIVE_DATA';
-
-export default RECEIVE_DATA;
